\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,PermSpace-method}
\title{Methods for function \code{initialize} in package \pkg{SOUP}: constructor}
\description{
  Methods for function \code{initialize} in package
  \pkg{SOUP}: constructor
}
\section{Methods}{
  \describe{ \item{\code{signature(object = "PermSpace")}}{
  constructor of the object (\code{initialize})} }
}

