\docType{methods}
\name{print}
\alias{print}
\alias{print,PValueMat-method}
\alias{print,PermSpace-method}
\alias{print,RankResults-method}
\alias{print,SoupObject-method}
\title{\describe{
\item{\code{signature(x = "PermSpace")}}{
     It prints the whole object on screen (mostly useful for
     external saving)}
}}
\description{
  \describe{ \item{\code{signature(x = "PermSpace")}}{ It
  prints the whole object on screen (mostly useful for
  external saving)} }

  \describe{ \item{\code{signature(x = "PValueMat")}}{ It
  prints the whole object on screen (mostly useful for
  external saving)} }

  \describe{ \item{\code{signature(x = "SoupObject")}}{ It
  prints the whole object on screen (mostly useful for
  external saving)} }
}

