\docType{methods}
\name{as.list}
\alias{as.list}
\alias{as.list,RankResults-method}
\title{Methods for function \code{as.list} in package \pkg{SOUP}}
\description{
  Methods for function \code{as.list} in package \pkg{SOUP}
}
\section{Methods}{
  \describe{ \item{\code{signature(object =
  "RankResults")}}{ conversion of the object
  (\code{RankResults}) into a \code{list}} }
}

