\docType{methods}
\name{show}
\alias{show}
\alias{show,PValueMat-method}
\alias{show,PermSpace-method}
\alias{show,RankResults-method}
\alias{show,SoupObject-method}
\title{\describe{
\item{\code{signature(object = "PermSpace")}}{
     Shows only the main information (on screen) for the object}
}}
\description{
  \describe{ \item{\code{signature(object = "PermSpace")}}{
  Shows only the main information (on screen) for the
  object} }

  Method show

  \describe{ \item{\code{signature(object = "PValueMat")}}{
  Shows only the main information (on screen) for the
  object} }

  \describe{ \item{\code{signature(object =
  "SoupObject")}}{ Shows only the main information (on
  screen) for the object} }
}

